+++ tests/testthat/test-summarize.R
@@ -66,17 +66,6 @@ test_that("can do group aggregation with no by", {
   expect_equal(tidytable_df, datatable_df)
 })
 
-# test_that("can do group aggregation with by list()", {
-#   df <- tidytable(x = 1:4, y = c("a","a","a","b"))
-#
-#   tidytable_df <- df %>%
-#     summarize.(avg_x = mean(x), by = list(y))
-#
-#   datatable_df <- df[, list(avg_x = mean(x)), by = y]
-#
-#   expect_equal(tidytable_df, datatable_df)
-# })
-
 test_that("by = list() causes an error", {
   df <- tidytable(x = 1:4, y = c("a","a","a","b"))
 
@@ -103,6 +92,20 @@ test_that("can do group aggregation with by c()", {
   datatable_df <- df[, list(avg_x = mean(x)), by = y]
 
   expect_equal(tidytable_df, datatable_df)
+})
+
+test_that("can use .sort & doesn't modify-by-reference", {
+  test_df <- tidytable(x = 1:3, y = c("b", "a", "a"), z = c("b", "a", "a"))
+
+  tidytable_df <- test_df %>%
+    summarize.(avg_x = mean(x),
+               .by = c(y, z),
+               .sort = TRUE)
+
+  datatable_df <- test_df[, list(avg_x = mean(x)), keyby = list(y, z)]
+
+  expect_equal(tidytable_df$avg_x, datatable_df$avg_x)
+  expect_named(test_df, c("x", "y", "z"))
 })
 
 test_that("can do group aggregation with by enhanced selection", {